refactoring : #40 - 패키지 이동 및 불필요 테스트 제거
