--- src/test/java/com/fesi/mukitlist/api/repository/usergathering/UserGatheringSpecificationsTest.java
@@ -1,20 +1,0 @@
-package com.fesi.mukitlist.api.repository.usergathering;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-import org.junit.jupiter.api.Test;
-
-class UserGatheringSpecificationsTest {
-
-	@Test
-	void byUser() {
-	}
-
-	@Test
-	void byCompleted() {
-	}
-
-	@Test
-	void byReviewed() {
-	}
-}
